+++ lib/analytics.dart
@@ -15,5 +15,4 @@
 export 'package:analytics/core/analytic_action.dart';
 export 'package:analytics/core/analytic_action_performer.dart';
 export 'package:analytics/core/analytic_service.dart';
-
-export 'package:analytics/impl/default_analytic_service.dart';
+export 'package:analytics/core/analytic_strategy.dart';