+++ lib/core/analytic_service.dart
@@ -12,10 +12,74 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import 'package:analytics/analytics.dart';
+import 'package:analytics/core/analytic_action.dart';
+import 'package:analytics/core/analytic_action_performer.dart';
+import 'package:analytics/core/analytic_strategy.dart';
+import 'package:analytics/utils/logger.dart';
+import 'package:logger/logger.dart';
 
-/// A unified entry point for several [AnalyticActionPerformer]s.
-// ignore: one_member_abstracts
-abstract class AnalyticService<A extends AnalyticAction> {
-  void performAction(A action);
+/// Class for instantiating analytic service.
+///
+/// WARNING! Do not use this class directly for analytic action sending.
+///
+/// This class provides a way to send analytic action using a set of
+/// analytic strategies. To use this class, you should create an instance of it with a set of
+/// analytic strategies, and then call the `performAction` method to send analytic action.
+class AnalyticService implements AnalyticActionPerformer {
+  /// Initialize analytic service without strategies.
+  AnalyticService.empty({Logger? logger})
+      : _strategies = <AnalyticStrategy>{},
+        _logger = logger ?? defaultLogger;
+
+  /// Initialize analytic service with strategies.
+  AnalyticService.withStrategies(this._strategies, {Logger? logger})
+      : _logger = logger ?? defaultLogger;
+  final Logger _logger;
+
+  final Set<AnalyticStrategy> _strategies;
+
+  @override
+  void performAction(AnalyticAction action) {
+    _getStrategiesByAction(action)
+        .forEach((strategy) => strategy.performAction(action));
+  }
+
+  List<AnalyticActionPerformer<AnalyticAction>> _getStrategiesByAction(
+    AnalyticAction event,
+  ) {
+    final properStrategies =
+        _strategies.where((strategies) => strategies.canHandle(event)).toList();
+    if (properStrategies.isEmpty) {
+      _logger.d(
+        'No action performer for action:'
+        ' ${event.runtimeType} in performers $_strategies',
+      );
+    }
+    return properStrategies;
+  }
+
+  /// Add strategy to analytic service.
+  void addStrategy(AnalyticStrategy strategy) {
+    _strategies.add(strategy);
+  }
+
+  /// Remove strategy from analytic service.
+  void removeStrategy(AnalyticStrategy strategy) {
+    _strategies.remove(strategy);
+  }
+
+  /// Remove all strategies from analytic service.
+  void clearStrategies() {
+    _strategies.clear();
+  }
+
+  /// Iterate over all strategies.
+  void forEach(void Function(AnalyticStrategy strategy) action) {
+    _strategies.forEach(action);
+  }
+
+  /// Check if a strategy is contained in the analytic service.
+  bool contains(AnalyticStrategy strategy) {
+    return _strategies.contains(strategy);
+  }
 }