+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 2.0.1
+
+Updated readme file
+
+## 2.0.0
+
+To use a specific analytics service, you need to define a strategy. Strategies are stored in the `AnalyticService` class, which has a `performAction` method that sends an event to specific analytics services by calling the strategies `performAction` method.
+The `AnalyticStrategy` and `AnalyticService` implement the `AnalyticActionPerformer`.
+
 ## 1.0.2
 
 * Internal improvement